Split up files a bit and rename
