+++ packages/sdk/wallets/wallet-keystore/src/index.ts
@@ -1,1 +1,2 @@
-export * from './local-keystore'
+export * from './keystore-wallet-wrapper'
+export * from './keystores'
